lax_numpy: move quantile-based functions to reductions.py
