+++ jax/numpy/__init__.py
@@ -107,6 +107,8 @@ from jax._src.numpy.lax_numpy import (
     float16 as float16,
     float32 as float32,
     float64 as float64,
+    float8_e4m3fn as float8_e4m3fn,
+    float8_e5m2 as float8_e5m2,
     float_ as float_,
     floating as floating,
     fmax as fmax,
@@ -166,7 +168,6 @@ from jax._src.numpy.lax_numpy import (
     logspace as logspace,
     mask_indices as mask_indices,
     matmul as matmul,
-    median as median,
     meshgrid as meshgrid,
     moveaxis as moveaxis,
     msort as msort,
@@ -175,9 +176,6 @@ from jax._src.numpy.lax_numpy import (
     nanargmax as nanargmax,
     nanargmin as nanargmin,
     argpartition as argpartition,
-    nanmedian as nanmedian,
-    nanpercentile as nanpercentile,
-    nanquantile as nanquantile,
     ndim as ndim,
     newaxis as newaxis,
     nonzero as nonzero,
@@ -189,14 +187,12 @@ from jax._src.numpy.lax_numpy import (
     packbits as packbits,
     pad as pad,
     partition as partition,
-    percentile as percentile,
     pi as pi,
     piecewise as piecewise,
     place as place,
     printoptions as printoptions,
     promote_types as promote_types,
     put as put,
-    quantile as quantile,
     ravel as ravel,
     ravel_multi_index as ravel_multi_index,
     repeat as repeat,
@@ -258,11 +254,6 @@ from jax._src.numpy.lax_numpy import (
     zeros_like as zeros_like,
 )
 
-from jax._src.numpy.lax_numpy import (
-  float8_e4m3fn,
-  float8_e5m2,
-)
-
 from jax._src.numpy.index_tricks import (
   c_ as c_,
   index_exp as index_exp,
@@ -298,19 +289,25 @@ from jax._src.numpy.reductions import (
     cumproduct as cumproduct,
     max as max,
     mean as mean,
+    median as median,
     min as min,
     nancumsum as nancumsum,
     nancumprod as nancumprod,
     nanmax as nanmax,
     nanmean as nanmean,
+    nanmedian as nanmedian,
     nanmin as nanmin,
+    nanpercentile as nanpercentile,
     nanprod as nanprod,
+    nanquantile as nanquantile,
     nanstd as nanstd,
     nansum as nansum,
     nanvar as nanvar,
+    percentile as percentile,
     prod as prod,
     product as product,
     ptp as ptp,
+    quantile as quantile,
     sometrue as sometrue,
     std as std,
     sum as sum,
